+++ apps/dashboard/src/@/api/universal-bridge/types.ts
@@ -1,0 +1,8 @@
+export type TokenMetadata = {
+  name: string;
+  symbol: string;
+  address: string;
+  decimals: number;
+  chainId: number;
+  iconUri?: string;
+};
